+++ TestBotDD/src/main/java/Commands/RolePost.java
@@ -4,6 +4,7 @@ import Commands_utilities.Command;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.ListMultimap;
 import com.google.common.collect.Multimap;
+import net.dv8tion.jda.api.Permission;
 import net.dv8tion.jda.api.events.message.MessageReceivedEvent;
 import org.apache.commons.collections4.MultiMap;
 
@@ -26,27 +27,29 @@ public class RolePost extends Command {// nie mam weny do nazw dzisiaj :V
 
     @Override
     protected void execute(MessageReceivedEvent e) {
-        String messageRaw = e.getMessage().getContentRaw();
-        String[] message = messageRaw.split(" ");
-        if(message.length>1){
+        if(e.getMember().hasPermission(Permission.valueOf("MANAGE_ROLES")) || RoleCommands.roleCheck(e.getMember().getRoles(),name)) {
+            String messageRaw = e.getMessage().getContentRaw();
+            String[] message = messageRaw.split(" ");
+            if (message.length > 1) {
 
-            //e.getChannel().sendMessage(String.join(" ", Arrays.copyOfRange(message, messageRaw.indexOf("content:"), message.length))).queue();
-            String substring = messageRaw.substring(messageRaw.indexOf("content: ") + 9);
-            e.getChannel().sendMessage(substring).queue();
-            e.getMessage().delete().queue();
+                //e.getChannel().sendMessage(String.join(" ", Arrays.copyOfRange(message, messageRaw.indexOf("content:"), message.length))).queue();
+                String substring = messageRaw.substring(messageRaw.indexOf("content: ") + 9);
+                e.getChannel().sendMessage(substring).queue();
+                e.getMessage().delete().queue();
 
-            int i =1;
-            while (!message[i].equals("content:")){
-                if(message[i].split("=").length==2){
-                    int j = i+1;
-                    String text = message[i];
-                    while(!(message[j].split("=").length==2)&&!message[j].equals("content:")){
-                        text = text+" "+ message[j];
-                        j++;
+                int i = 1;
+                while (!message[i].equals("content:")) {
+                    if (message[i].split("=").length == 2) {
+                        int j = i + 1;
+                        String text = message[i];
+                        while (!(message[j].split("=").length == 2) && !message[j].equals("content:")) {
+                            text = text + " " + message[j];
+                            j++;
+                        }
+                        postMap.put(substring, text);
                     }
-                    postMap.put(substring,text);
+                    i++;
                 }
-                i++;
             }
         }
     }
